Minor changes to CSS and navbar
Amended break points for media queries to ensure that the site was mobile friendly.  I also changed the subscribe button to a colour with more contrast. While checking the site responsiveness, I also noticed the the navbar need to be changed to a collapsible navbar at a wider screen px.