Minor changes to opportunities and subscribe page
I made changes to CSS and bootstrap on the opportunities page, this was to ensure that the page flowed better and was more visually appealing. I added a background image to the subscription page as i thought that the page was in need of something more, considered adding more content but it wouldn't of added any value. Minor changes to readme file, more will be added later on.