Minor Update to meta elements on pages within the site, minor css changes to colors and added more content to readme file
Added meta element to each page which included description, keywords and author. Minor change to css included streamlining the colors in css file so that they are all displayed the same and added more content to the readme file, this included more content in the credits section and minor additions to other sections including wireframe image added.