readme update, added meta element to the head and minor change to form on subscribe page
I made changes to the markup language as there were errors present in previous code, added content to the feature section in the readme file including screenshots of the site. added meta elements to the head which include description, keywords and author, added 'target_blank' to the form so that when submitted, it opens in a new tab.